--- tests/mcp.integration.test.ts
@@ -1,3032 +1,0 @@
-import { afterEach, beforeAll, describe, expect, it } from "vitest";
-import { mkdtemp, mkdir, readFile, readdir, rm, stat, writeFile } from "fs/promises";
-import os from "os";
-import path from "path";
-import { spawn } from "child_process";
-import http from "http";
-import { fileURLToPath } from "url";
-import { promisify } from "util";
-import { execFile } from "child_process";
-
-import {
-  ConsolidateResultSchema,
-  MemoryGraphResultSchema,
-  MigrationExecuteResultSchema,
-  MigrationListResultSchema,
-  OrientationNoteSchema,
-  ProjectSummaryResultSchema,
-  RecallResultSchema,
-  SyncResultSchema,
-} from "../src/structured-content.js";
-
-const repoRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
-const execFileAsync = promisify(execFile);
-const builtEntryPoint = path.join(repoRoot, "build", "index.js");
-
-const tempDirs: string[] = [];
-
-async function initTestRepo(repoDir: string, branch = "feature/test"): Promise<void> {
-  await execFileAsync("git", ["init"], { cwd: repoDir });
-  await execFileAsync("git", ["checkout", "-B", branch], { cwd: repoDir });
-  await execFileAsync("git", ["config", "user.name", "Test User"], { cwd: repoDir });
-  await execFileAsync("git", ["config", "user.email", "test@example.com"], { cwd: repoDir });
-}
-
-async function initTestVaultRepo(vaultDir: string): Promise<void> {
-  await execFileAsync("git", ["init"], { cwd: vaultDir });
-  await execFileAsync("git", ["config", "user.name", "Test User"], { cwd: vaultDir });
-  await execFileAsync("git", ["config", "user.email", "test@example.com"], { cwd: vaultDir });
-}
-
-afterEach(async () => {
-  await Promise.all(tempDirs.splice(0).map((dir) => rm(dir, { recursive: true, force: true })));
-});
-
-beforeAll(async () => {
-  await execFileAsync("npm", ["run", "build"], { cwd: repoRoot });
-}, 120000);
-
-describe("local MCP script", () => {
-  it("exposes workflow-hint prompt as an imperative decision protocol", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-
-    const promptText = await callLocalMcpPrompt(vaultDir, "mnemonic-workflow-hint");
-
-    expect(promptText).toContain("Avoid duplicate memories.");
-    expect(promptText).toContain("REQUIRES: Before `remember`, call `recall` or `list` first.");
-    expect(promptText).toContain("If `recall` or `list` returns a plausible match, call `get` before deciding whether to `update` or `remember`.");
-    expect(promptText).toContain("When unsure, prefer `recall` over `remember`.");
-    expect(promptText).toContain("Bad: call `remember` immediately because the user said 'remember'.");
-  }, 15000);
-
-  it("surfaces prerequisite-first workflow wording in phase-aware tool descriptions", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-
-    const tools = await listLocalMcpTools(vaultDir);
-    const byName = new Map(tools.map((tool) => [tool.name, tool.description ?? ""]));
-
-    expect(byName.get("remember")).toContain("REQUIRES: Call `recall` or `list` first to check whether this memory already exists.");
-    expect(byName.get("get")).toContain("Use after `recall`, `list`, or `recent_memories` when you need the full note content.");
-    expect(byName.get("update")).toContain("Use after `recall` + `get` when an existing memory should be refined instead of creating a duplicate.");
-    expect(byName.get("relate")).toContain("Use after you have identified the exact memories to connect.");
-    expect(byName.get("consolidate")).toContain("Use after `recall`, `list`, or `memory_graph` shows overlap that should be merged or cleaned up.");
-    expect(byName.get("move_memory")).toContain("Use after `where_is_memory` or `get` confirms a memory is stored in the wrong place.");
-    expect(byName.get("discover_tags")).toContain("Suggest canonical tags for a specific note before `remember` when tag choice is ambiguous.");
-  }, 15000);
-
-  it("supports global remember and forget with git disabled", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Integration dogfood note",
-        content: "Temporary integration-test note created through the local MCP script.",
-        tags: ["integration", "dogfood"],
-        summary: "Create integration dogfood note with git disabled",
-        scope: "global",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-      const notePath = path.join(vaultDir, "notes", `${noteId}.md`);
-
-      await expect(stat(notePath)).resolves.toBeDefined();
-      await expect(readFile(notePath, "utf-8")).resolves.toContain("Integration dogfood note");
-
-      const forgetText = await callLocalMcp(vaultDir, "forget", { id: noteId }, embeddingServer.url);
-      expect(forgetText).toContain(`Forgotten '${noteId}'`);
-      await expect(stat(notePath)).rejects.toThrow();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("returns structured persistence details for remember without extra verification calls", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const response = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Persistence status remember test",
-        content: "Verify remember returns persistence metadata.",
-        tags: ["integration"],
-        scope: "global",
-        summary: "Create note and inspect persistence result",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(response.text);
-      const structured = response.structuredContent;
-      expect(structured?.["action"]).toBe("remembered");
-      const persistence = structured?.["persistence"] as Record<string, unknown>;
-      expect(persistence?.["notePath"]).toBe(path.join(vaultDir, "notes", `${noteId}.md`));
-      expect(persistence?.["embeddingPath"]).toBe(path.join(vaultDir, "embeddings", `${noteId}.json`));
-      expect((persistence?.["embedding"] as Record<string, unknown>)?.["status"]).toBe("written");
-      const git = persistence?.["git"] as Record<string, unknown>;
-      expect(git?.["commit"]).toBe("skipped");
-      expect(git?.["push"]).toBe("skipped");
-      expect(git?.["commitMessage"]).toBe("remember: Persistence status remember test");
-      expect(String(git?.["commitBody"] ?? "")).toContain("Create note and inspect persistence result");
-      expect(persistence?.["durability"]).toBe("local-only");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("reports embedding skip reasons in structured persistence when Ollama is unavailable", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-
-    const response = await callLocalMcpResponse(vaultDir, "remember", {
-      title: "Persistence status embedding failure",
-      content: "This note should survive even if embedding fails.",
-      tags: ["integration"],
-      scope: "global",
-      summary: "Create note with embedding failure",
-    }, { ollamaUrl: "http://127.0.0.1:9" });
-
-    const structured = response.structuredContent;
-    expect(structured?.["action"]).toBe("remembered");
-    const persistence = structured?.["persistence"] as Record<string, unknown>;
-    const embedding = persistence?.["embedding"] as Record<string, unknown>;
-    expect(embedding?.["status"]).toBe("skipped");
-    expect(String(embedding?.["reason"] ?? "")).not.toBe("");
-    expect(persistence?.["durability"]).toBe("local-only");
-  }, 15000);
-
-  it("returns deterministic retry metadata when git.add fails after remember mutation", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      await execFileAsync("git", ["init"], { cwd: vaultDir });
-      // Create a persistent index.lock - this will cause git.add() to fail
-      // and our retry logic will exhaust 3 attempts before giving up.
-      await writeFile(path.join(vaultDir, ".git", "index.lock"), "locked\n", "utf-8");
-
-      const response = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Retry contract remember test",
-        content: "This should write the note even if commit fails.",
-        tags: ["integration", "retry"],
-        scope: "global",
-        summary: "Capture deterministic retry metadata after commit failure",
-      }, { ollamaUrl: embeddingServer.url, disableGit: false });
-
-      const noteId = extractRememberedId(response.text);
-      await expect(stat(path.join(vaultDir, "notes", `${noteId}.md`))).resolves.toBeDefined();
-
-      const persistence = response.structuredContent?.["persistence"] as Record<string, unknown>;
-      const git = persistence?.["git"] as Record<string, unknown>;
-      expect(git?.["commit"]).toBe("failed");
-      expect(git?.["commitOperation"]).toBe("add");
-      expect(git?.["push"]).toBe("skipped");
-      expect(String(git?.["commitError"] ?? "")).toContain("index.lock");
-      expect(response.text).toContain("Git add error:");
-      expect(response.text).toContain("Retry: safe");
-
-      const retry = persistence?.["retry"] as Record<string, unknown>;
-      expect(retry?.["mutationApplied"]).toBe(true);
-      expect(retry?.["retrySafe"]).toBe(true);
-      const attemptedCommit = retry?.["attemptedCommit"] as Record<string, unknown>;
-      expect(attemptedCommit?.["message"]).toBe("remember: Retry contract remember test");
-      expect(attemptedCommit?.["vault"]).toBe("main-vault");
-      expect(attemptedCommit?.["operation"]).toBe("add");
-      expect((attemptedCommit?.["files"] as string[])).toHaveLength(1);
-      expect(String(attemptedCommit?.["error"] ?? "")).toContain("index.lock");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("supports overriding project identity to use upstream", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-    await execFileAsync("git", ["remote", "add", "origin", "git@github.com:user/myapp-fork.git"], { cwd: repoDir });
-    await execFileAsync("git", ["remote", "add", "upstream", "git@github.com:acme/myapp.git"], { cwd: repoDir });
-
-    const before = await callLocalMcp(vaultDir, "get_project_identity", { cwd: repoDir });
-    expect(before).toContain("`github-com-user-myapp-fork`");
-    expect(before).toContain("**remote:** origin");
-
-    const setResult = await callLocalMcp(vaultDir, "set_project_identity", {
-      cwd: repoDir,
-      remoteName: "upstream",
-    });
-    expect(setResult).toContain("default=`github-com-user-myapp-fork`");
-    expect(setResult).toContain("effective=`github-com-acme-myapp`");
-
-    const after = await callLocalMcp(vaultDir, "get_project_identity", { cwd: repoDir });
-    expect(after).toContain("`github-com-acme-myapp`");
-    expect(after).toContain("**remote:** upstream");
-    expect(after).toContain("**default id:** `github-com-user-myapp-fork`");
-  }, 15000);
-
-  it("returns retry metadata when policy commit fails after config mutation", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-    await execFileAsync("git", ["init"], { cwd: vaultDir });
-    await writeFile(path.join(vaultDir, ".git", "index.lock"), "locked\n", "utf-8");
-
-    const response = await callLocalMcpResponse(vaultDir, "set_project_memory_policy", {
-      cwd: repoDir,
-      defaultScope: "global",
-    }, { disableGit: false });
-
-    expect(response.text).toContain("defaultScope=global");
-    expect(response.text).toContain("Git add error:");
-    expect(response.text).toContain("Retry: safe");
-    const retry = response.structuredContent?.["retry"] as Record<string, unknown>;
-    expect(retry?.["mutationApplied"]).toBe(true);
-    const attempted = retry?.["attemptedCommit"] as Record<string, unknown>;
-    expect(attempted?.["message"]).toContain("policy:");
-    expect(attempted?.["vault"]).toBe("main-vault");
-    expect(String(attempted?.["error"] ?? "")).toContain("index.lock");
-  }, 15000);
-
-  it("surfaces retry guidance in text when relate commit fails", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      await initTestVaultRepo(vaultDir);
-
-      const first = await callLocalMcp(vaultDir, "remember", {
-        title: "Relate retry first",
-        content: "First note for relate retry test.",
-        tags: ["integration", "retry"],
-        scope: "global",
-        summary: "Create first note for relate retry test",
-      }, { ollamaUrl: embeddingServer.url, disableGit: false });
-
-      const second = await callLocalMcp(vaultDir, "remember", {
-        title: "Relate retry second",
-        content: "Second note for relate retry test.",
-        tags: ["integration", "retry"],
-        scope: "global",
-        summary: "Create second note for relate retry test",
-      }, { ollamaUrl: embeddingServer.url, disableGit: false });
-
-      const firstId = extractRememberedId(first);
-      const secondId = extractRememberedId(second);
-
-      await writeFile(path.join(vaultDir, ".git", "index.lock"), "locked\n", "utf-8");
-
-      const response = await callLocalMcpResponse(vaultDir, "relate", {
-        fromId: firstId,
-        toId: secondId,
-      }, { ollamaUrl: embeddingServer.url, disableGit: false });
-
-      expect(response.text).toContain(`Linked \`${firstId}\` ↔ \`${secondId}\` (related-to)`);
-      expect(response.text).toContain("Git add error:");
-      expect(response.text).toContain("Retry: safe");
-
-      const retry = response.structuredContent?.["retry"] as Record<string, unknown>;
-      expect(retry?.["mutationApplied"]).toBe(true);
-      expect(retry?.["retrySafe"]).toBe(true);
-      const attempted = retry?.["attemptedCommit"] as Record<string, unknown>;
-      expect(attempted?.["message"]).toContain("relate:");
-      expect(attempted?.["vault"]).toBe("main-vault");
-      expect(String(attempted?.["error"] ?? "")).toContain("index.lock");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("applies project memory policy end-to-end for remember routing", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const setGlobal = await callLocalMcpResponse(vaultDir, "set_project_memory_policy", {
-        cwd: repoDir,
-        defaultScope: "global",
-      }, embeddingServer.url);
-      expect(setGlobal.text).toContain("defaultScope=global");
-
-      const getGlobal = await callLocalMcpResponse(vaultDir, "get_project_memory_policy", { cwd: repoDir }, embeddingServer.url);
-      expect(getGlobal.text).toContain("defaultScope=global");
-      expect(getGlobal.structuredContent?.["defaultScope"]).toBe("global");
-
-      const rememberedGlobal = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Policy global default note",
-        content: "Should land in the main vault while keeping project association.",
-        tags: ["integration", "policy"],
-        summary: "Use global policy default for remember routing",
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const globalId = extractRememberedId(rememberedGlobal.text);
-      expect(rememberedGlobal.structuredContent?.["scope"]).toBe("global");
-      expect(rememberedGlobal.structuredContent?.["vault"]).toBe("main-vault");
-      await expect(stat(path.join(vaultDir, "notes", `${globalId}.md`))).resolves.toBeDefined();
-
-      const setProject = await callLocalMcpResponse(vaultDir, "set_project_memory_policy", {
-        cwd: repoDir,
-        defaultScope: "project",
-      }, embeddingServer.url);
-      expect(setProject.text).toContain("defaultScope=project");
-
-      const rememberedProject = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Policy project default note",
-        content: "Should land in the project vault when scope is omitted.",
-        tags: ["integration", "policy"],
-        summary: "Use project policy default for remember routing",
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const projectId = extractRememberedId(rememberedProject.text);
-      expect(rememberedProject.structuredContent?.["scope"]).toBe("project");
-      expect(rememberedProject.structuredContent?.["vault"]).toBe("project-vault");
-      await expect(stat(path.join(repoDir, ".mnemonic", "notes", `${projectId}.md`))).resolves.toBeDefined();
-
-      const setAsk = await callLocalMcpResponse(vaultDir, "set_project_memory_policy", {
-        cwd: repoDir,
-        defaultScope: "ask",
-      }, embeddingServer.url);
-      expect(setAsk.text).toContain("defaultScope=ask");
-
-      const askRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Policy ask note",
-        content: "Should not be written until scope is explicit.",
-        tags: ["integration", "policy"],
-        summary: "Require explicit scope when policy is ask",
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      expect(askRemember).toContain("always ask");
-      expect(askRemember).toContain("scope: \"project\"");
-      expect(askRemember).toContain("scope: \"global\"");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("applies protected branch policy for project-vault remember writes", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-    await execFileAsync("git", ["checkout", "-B", "main"], { cwd: repoDir });
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const explicitScopeBlocked = await callLocalMcp(vaultDir, "remember", {
-        title: "Protected branch explicit scope blocked note",
-        content: "Explicit project scope should still respect protected-branch policy.",
-        tags: ["integration", "protected-branch"],
-        summary: "Block explicit project scope remember on protected branch",
-        cwd: repoDir,
-        scope: "project",
-      }, embeddingServer.url);
-
-      expect(explicitScopeBlocked).toContain("Protected branch check");
-      expect(explicitScopeBlocked).toContain("allowProtectedBranch: true");
-
-      const bootstrapRemember = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Protected branch bootstrap note",
-        content: "Bootstraps project vault adoption with explicit scope.",
-        tags: ["integration", "protected-branch"],
-        summary: "Bootstrap project vault for protected branch policy test",
-        cwd: repoDir,
-        scope: "project",
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-      const bootstrapId = extractRememberedId(bootstrapRemember.text);
-      await expect(stat(path.join(repoDir, ".mnemonic", "notes", `${bootstrapId}.md`))).resolves.toBeDefined();
-
-      const protectedAsk = await callLocalMcp(vaultDir, "remember", {
-        title: "Protected branch ask note",
-        content: "Should prompt for branch policy before writing.",
-        tags: ["integration", "protected-branch"],
-        summary: "Prompt for protected branch commit behavior",
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      expect(protectedAsk).toContain("Protected branch check");
-      expect(protectedAsk).toContain("allowProtectedBranch: true");
-      expect(protectedAsk).toContain("protectedBranchBehavior");
-
-      const setAllow = await callLocalMcpResponse(vaultDir, "set_project_memory_policy", {
-        cwd: repoDir,
-        protectedBranchBehavior: "allow",
-      }, embeddingServer.url);
-      expect(setAllow.text).toContain("protectedBranchBehavior=allow");
-      expect(setAllow.structuredContent?.["protectedBranchBehavior"]).toBe("allow");
-
-      const allowedRemember = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Protected branch allowed note",
-        content: "Should write and commit on protected branch after allow policy.",
-        tags: ["integration", "protected-branch"],
-        summary: "Allow protected branch remember commits for this project",
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const allowedId = extractRememberedId(allowedRemember.text);
-      await expect(stat(path.join(repoDir, ".mnemonic", "notes", `${allowedId}.md`))).resolves.toBeDefined();
-
-      const setBlock = await callLocalMcpResponse(vaultDir, "set_project_memory_policy", {
-        cwd: repoDir,
-        protectedBranchBehavior: "block",
-      }, embeddingServer.url);
-      expect(setBlock.text).toContain("protectedBranchBehavior=block");
-
-      const protectedBlock = await callLocalMcp(vaultDir, "remember", {
-        title: "Protected branch blocked note",
-        content: "Should be blocked when policy says block.",
-        tags: ["integration", "protected-branch"],
-        summary: "Block protected branch remember commits for this project",
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      expect(protectedBlock).toContain("Auto-commit blocked");
-
-      const overrideRemember = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Protected branch override note",
-        content: "One-time override should bypass block policy.",
-        tags: ["integration", "protected-branch"],
-        summary: "Use one-time protected branch override for remember",
-        cwd: repoDir,
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-
-      const overrideId = extractRememberedId(overrideRemember.text);
-      await expect(stat(path.join(repoDir, ".mnemonic", "notes", `${overrideId}.md`))).resolves.toBeDefined();
-
-      const protectedUpdateAsk = await callLocalMcp(vaultDir, "update", {
-        id: overrideId,
-        content: "Update should ask for protected-branch override.",
-        cwd: repoDir,
-      }, embeddingServer.url);
-      expect(protectedUpdateAsk).toContain("Auto-commit blocked");
-      expect(protectedUpdateAsk).toContain("`update`");
-
-      const protectedUpdateOverride = await callLocalMcp(vaultDir, "update", {
-        id: overrideId,
-        content: "Update with one-time override should proceed.",
-        cwd: repoDir,
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-      expect(protectedUpdateOverride).toContain(`Updated memory '${overrideId}'`);
-
-      const projectForgetRemember = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Protected branch forget note",
-        content: "Used to verify protected branch checks for forget.",
-        tags: ["integration", "protected-branch"],
-        summary: "Create note for protected branch forget test",
-        cwd: repoDir,
-        scope: "project",
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-      const forgetId = extractRememberedId(projectForgetRemember.text);
-
-      const protectedForgetAsk = await callLocalMcp(vaultDir, "forget", {
-        id: forgetId,
-        cwd: repoDir,
-      }, embeddingServer.url);
-      expect(protectedForgetAsk).toContain("Auto-commit blocked");
-      expect(protectedForgetAsk).toContain("`forget`");
-
-      const protectedForgetOverride = await callLocalMcp(vaultDir, "forget", {
-        id: forgetId,
-        cwd: repoDir,
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-      expect(protectedForgetOverride).toContain(`Forgotten '${forgetId}'`);
-
-      const moveRemember = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Protected branch move source",
-        content: "Created as global note for move protected-branch test.",
-        tags: ["integration", "protected-branch"],
-        summary: "Create global note for protected branch move test",
-        scope: "global",
-      }, embeddingServer.url);
-      const moveId = extractRememberedId(moveRemember.text);
-
-      const protectedMoveAsk = await callLocalMcp(vaultDir, "move_memory", {
-        id: moveId,
-        target: "project-vault",
-        cwd: repoDir,
-      }, embeddingServer.url);
-      expect(protectedMoveAsk).toContain("Auto-commit blocked");
-      expect(protectedMoveAsk).toContain("`move_memory`");
-
-      const protectedMoveOverride = await callLocalMcp(vaultDir, "move_memory", {
-        id: moveId,
-        target: "project-vault",
-        cwd: repoDir,
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-      expect(protectedMoveOverride).toContain(`Moved '${moveId}'`);
-
-      const consolidateA = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Protected branch consolidate A",
-        content: "First note for consolidate policy coverage.",
-        tags: ["integration", "protected-branch", "consolidate"],
-        summary: "Create first consolidate source note",
-        cwd: repoDir,
-        scope: "project",
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-      const consolidateAId = extractRememberedId(consolidateA.text);
-
-      const consolidateB = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Protected branch consolidate B",
-        content: "Second note for consolidate policy coverage.",
-        tags: ["integration", "protected-branch", "consolidate"],
-        summary: "Create second consolidate source note",
-        cwd: repoDir,
-        scope: "project",
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-      const consolidateBId = extractRememberedId(consolidateB.text);
-
-      const protectedConsolidateAsk = await callLocalMcp(vaultDir, "consolidate", {
-        cwd: repoDir,
-        strategy: "execute-merge",
-        mergePlan: {
-          sourceIds: [consolidateAId, consolidateBId],
-          targetTitle: "Protected branch consolidate target",
-        },
-      }, embeddingServer.url);
-      expect(protectedConsolidateAsk).toContain("Auto-commit blocked");
-      expect(protectedConsolidateAsk).toContain("`consolidate`");
-
-      const protectedConsolidateOverride = await callLocalMcp(vaultDir, "consolidate", {
-        cwd: repoDir,
-        strategy: "execute-merge",
-        mergePlan: {
-          sourceIds: [consolidateAId, consolidateBId],
-          targetTitle: "Protected branch consolidate target",
-        },
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-      expect(protectedConsolidateOverride).toContain("Consolidated 2 notes");
-
-      const protectedPruneAsk = await callLocalMcp(vaultDir, "consolidate", {
-        cwd: repoDir,
-        strategy: "prune-superseded",
-        mode: "delete",
-      }, embeddingServer.url);
-      expect(protectedPruneAsk).toContain("Auto-commit blocked");
-      expect(protectedPruneAsk).toContain("`consolidate`");
-
-      const protectedPruneOverride = await callLocalMcp(vaultDir, "consolidate", {
-        cwd: repoDir,
-        strategy: "prune-superseded",
-        mode: "delete",
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-      expect(protectedPruneOverride).toContain("Pruned");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 35000);
-
-  it("skips auto-push for project-vault mutations by default", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const remoteDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-remote-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, remoteDir, repoDir);
-
-    await execFileAsync("git", ["init", "--bare"], { cwd: remoteDir });
-    await initTestRepo(repoDir);
-    await execFileAsync("git", ["remote", "add", "origin", remoteDir], { cwd: repoDir });
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const response = await callLocalMcpResponse(
-        vaultDir,
-        "remember",
-        {
-          title: "Project mutation push mode test",
-          content: "Project-vault writes should commit locally without pushing unpublished branches by default.",
-          tags: ["integration"],
-          summary: "Avoid auto-push for unpublished project branches",
-          cwd: repoDir,
-          scope: "project",
-        },
-        { ollamaUrl: embeddingServer.url, disableGit: false },
-      );
-
-      expect(response.text).toContain("Persistence: embedding written | git committed");
-      const structured = response.structuredContent as Record<string, unknown>;
-      const persistence = structured?.["persistence"] as Record<string, unknown>;
-      const git = persistence?.["git"] as Record<string, unknown>;
-      expect(persistence?.["durability"]).toBe("committed");
-      expect(git?.["push"]).toBe("skipped");
-      expect(git?.["pushReason"]).toBe("auto-push-disabled");
-
-      const noteId = structured?.["id"] as string;
-      await expect(stat(path.join(repoDir, ".mnemonic", "notes", `${noteId}.md`))).resolves.toBeDefined();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("rewrites project metadata when moving a global note into a project vault", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Unscoped move test",
-        content: "Created without project context so it starts as a global note.",
-        tags: ["integration"],
-        summary: "Create unscoped note for move metadata rewrite test",
-        scope: "global",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-      const moveText = await callLocalMcp(vaultDir, "move_memory", {
-        id: noteId,
-        target: "project-vault",
-        cwd: repoDir,
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-
-      expect(moveText).toContain("Project association is now");
-      expect(moveText).toContain("(");
-
-      const movedNote = await readFile(path.join(repoDir, ".mnemonic", "notes", `${noteId}.md`), "utf-8");
-      expect(movedNote).toContain("projectName:");
-      expect(movedNote).toContain("project:");
-      expect(movedNote).toContain("mnemonic-mcp-project-");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("preserves project association when moving a project note into the main vault", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Project move out test",
-        content: "Created with project context so it should remain project-associated when moved to the main vault.",
-        tags: ["integration"],
-        summary: "Create project note for move-out behavior test",
-        cwd: repoDir,
-        scope: "project",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-      const moveText = await callLocalMcp(vaultDir, "move_memory", {
-        id: noteId,
-        target: "main-vault",
-        cwd: repoDir,
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-
-      expect(moveText).toContain("Project association remains");
-      expect(moveText).toContain("(");
-
-      const movedNote = await readFile(path.join(vaultDir, "notes", `${noteId}.md`), "utf-8");
-      expect(movedNote).toContain("projectName:");
-      expect(movedNote).toContain("project:");
-      expect(movedNote).toContain("mnemonic-mcp-project-");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("keeps note visibility coherent when moving a note from main to project and back", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Round trip move test",
-        content: "Start in the main vault and move through both storage locations.",
-        tags: ["integration", "move"],
-        summary: "Create note for move round-trip visibility test",
-        scope: "global",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-
-      const moveToProject = await callLocalMcpResponse(vaultDir, "move_memory", {
-        id: noteId,
-        target: "project-vault",
-        cwd: repoDir,
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-      expect(moveToProject.text).toContain("Project association is now");
-
-      const moveBackToMain = await callLocalMcpResponse(vaultDir, "move_memory", {
-        id: noteId,
-        target: "main-vault",
-        cwd: repoDir,
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-      expect(moveBackToMain.text).toContain("Project association remains");
-
-      await expect(stat(path.join(vaultDir, "notes", `${noteId}.md`))).resolves.toBeDefined();
-      await expect(stat(path.join(repoDir, ".mnemonic", "notes", `${noteId}.md`))).rejects.toThrow();
-
-      const listed = await callLocalMcpResponse(vaultDir, "list", {
-        cwd: repoDir,
-        scope: "project",
-        storedIn: "main-vault",
-        includeStorage: true,
-        includeUpdated: true,
-      }, embeddingServer.url);
-
-      expect(listed.text).toContain("Round trip move test");
-      expect(listed.text).toContain("stored=main-vault");
-      expect(listed.structuredContent?.["count"]).toBe(1);
-      const notes = listed.structuredContent?.["notes"] as Array<Record<string, unknown>>;
-      expect(notes[0]?.["id"]).toBe(noteId);
-      expect(notes[0]?.["vault"]).toBe("main-vault");
-      expect(notes[0]?.["project"]).toBeTruthy();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("shows consistent cross-vault results for list recent_memories and project_memory_summary", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const privateProjectRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Private project memory",
-        content: "Stored in main vault but associated with the current project.",
-        tags: ["integration", "cross-vault"],
-        summary: "Create private project memory for visibility test",
-        cwd: repoDir,
-        scope: "global",
-      }, embeddingServer.url);
-      const privateProjectId = extractRememberedId(privateProjectRemember);
-
-      const sharedProjectRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Shared project memory",
-        content: "Stored in the project vault for the current repo.",
-        tags: ["integration", "cross-vault"],
-        summary: "Create shared project memory for visibility test",
-        cwd: repoDir,
-        scope: "project",
-      }, embeddingServer.url);
-      const sharedProjectId = extractRememberedId(sharedProjectRemember);
-
-      const globalRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Unscoped global memory",
-        content: "Stored in main vault without project association.",
-        tags: ["integration", "cross-vault"],
-        summary: "Create unscoped global memory for visibility test",
-        scope: "global",
-      }, embeddingServer.url);
-      const globalId = extractRememberedId(globalRemember);
-
-      const listed = await callLocalMcpResponse(vaultDir, "list", {
-        cwd: repoDir,
-        scope: "all",
-        storedIn: "any",
-        tags: ["integration", "cross-vault"],
-        includeStorage: true,
-        includeUpdated: true,
-      }, embeddingServer.url);
-
-      expect(listed.structuredContent?.["count"]).toBe(3);
-      const listedNotes = listed.structuredContent?.["notes"] as Array<Record<string, unknown>>;
-      expect(listedNotes.map((note) => note["id"])).toEqual([privateProjectId, sharedProjectId, globalId]);
-      expect(listed.text).toContain("stored=project-vault");
-      expect(listed.text).toContain("stored=main-vault");
-
-      const recent = await callLocalMcpResponse(vaultDir, "recent_memories", {
-        cwd: repoDir,
-        scope: "project",
-        storedIn: "any",
-        limit: 5,
-        includePreview: false,
-        includeStorage: true,
-      }, embeddingServer.url);
-
-      expect(recent.structuredContent?.["count"]).toBe(2);
-      const recentNotes = recent.structuredContent?.["notes"] as Array<Record<string, unknown>>;
-      expect(recentNotes.map((note) => note["id"])).toEqual([sharedProjectId, privateProjectId]);
-      expect(recent.text).not.toContain("Unscoped global memory");
-
-      const summary = await callLocalMcpResponse(vaultDir, "project_memory_summary", {
-        cwd: repoDir,
-        recentLimit: 5,
-      }, embeddingServer.url);
-
-      // Project-scoped counts exclude unscoped global notes
-      const summaryNotes = summary.structuredContent?.["notes"] as Record<string, unknown>;
-      expect(summaryNotes?.["total"]).toBe(2); // only project-scoped notes
-      expect(summaryNotes?.["projectVault"]).toBe(1); // sharedProject
-      expect(summaryNotes?.["mainVault"]).toBe(1); // privateProject (in main vault with project association)
-      expect(summaryNotes?.["privateProject"]).toBe(1);
-      expect(summary.text).toContain("private project memories: 1");
-      expect(summary.text).not.toContain("Unscoped global memory"); // unscoped global should not appear
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("returns empty result when project has no memories but global vault does", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      // Create unscoped global memory (no project association)
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Unscoped global memory",
-        content: "Stored in main vault without project association.",
-        tags: ["integration", "empty-project"],
-        summary: "Create unscoped global memory",
-        scope: "global",
-      }, embeddingServer.url);
-
-      // Project has no memories, only global unscoped notes exist
-      const summary = await callLocalMcpResponse(vaultDir, "project_memory_summary", {
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      expect(summary.text).toContain("No memories found");
-      const summaryNotes = summary.structuredContent?.["notes"] as Record<string, unknown>;
-      expect(summaryNotes?.["total"]).toBe(0);
-      expect(summaryNotes?.["projectVault"]).toBe(0);
-      expect(summaryNotes?.["mainVault"]).toBe(0);
-      
-      const orientation = summary.structuredContent?.["orientation"] as Record<string, unknown>;
-      expect(orientation?.["primaryEntry"]).toEqual({
-        id: "",
-        title: "No notes",
-        rationale: "Empty project vault",
-      });
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("ranks tagged anchors by score not alphabetical order", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      // Create two tagged anchors with different centrality (permanent for anchor scoring)
-      // Alpha-first but lower centrality
-      const alphaRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "AAA Tagged Anchor",
-        content: "Alphabetically first but lower centrality.",
-        tags: ["anchor", "integration"],
-        summary: "Create alphabetically first tagged anchor",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-      const alphaId = extractRememberedId(alphaRemember);
-
-      // Create notes to link to (permanent for anchor scoring)
-      const note1 = await callLocalMcp(vaultDir, "remember", {
-        title: "Note to link",
-        content: "A note.",
-        tags: ["integration"],
-        summary: "Create note to link",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-      const note1Id = extractRememberedId(note1);
-
-      // Beta-later but higher centrality
-      const betaRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "ZZZ Tagged Anchor",
-        content: "Alphabetically last but higher centrality.",
-        tags: ["anchor", "integration"],
-        summary: "Create alphabetically last tagged anchor with more connections",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-      const betaId = extractRememberedId(betaRemember);
-
-      // Link beta to more notes (higher centrality)
-      await callLocalMcp(vaultDir, "relate", {
-        fromId: betaId,
-        toId: note1Id,
-        type: "related-to",
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const summary = await callLocalMcpResponse(vaultDir, "project_memory_summary", {
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const anchors = summary.structuredContent?.["anchors"] as Array<Record<string, unknown>>;
-      expect(anchors.length).toBeGreaterThan(0);
-      
-      // Beta should come before alpha despite alphabetical order
-      const betaIndex = anchors.findIndex(a => a["id"] === betaId);
-      const alphaIndex = anchors.findIndex(a => a["id"] === alphaId);
-      
-      // Both should be present
-      expect(betaIndex).toBeGreaterThanOrEqual(0);
-      expect(alphaIndex).toBeGreaterThanOrEqual(0);
-      
-      // Beta (higher centrality) should rank before alpha (lower centrality)
-      expect(betaIndex).toBeLessThan(alphaIndex);
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("orientation primaryEntry is best anchor by score", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      // Create a very connected anchor (permanent for anchor scoring)
-      const hubRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Hub Note",
-        content: "Central hub with many connections.",
-        tags: ["integration"],
-        summary: "Create hub note",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-      const hubId = extractRememberedId(hubRemember);
-
-      // Create peripheral notes (permanent for anchor scoring)
-      const periph1 = await callLocalMcp(vaultDir, "remember", {
-        title: "Peripheral 1",
-        content: "A peripheral note.",
-        tags: ["integration"],
-        summary: "Create peripheral note 1",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-      const periph1Id = extractRememberedId(periph1);
-
-      const periph2 = await callLocalMcp(vaultDir, "remember", {
-        title: "Peripheral 2",
-        content: "Another peripheral note.",
-        tags: ["integration"],
-        summary: "Create peripheral note 2",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-      const periph2Id = extractRememberedId(periph2);
-
-      // Connect hub to peripherals (need cwd for project-scoped notes)
-      await callLocalMcp(vaultDir, "relate", {
-        fromId: hubId,
-        toId: periph1Id,
-        type: "related-to",
-        cwd: repoDir,
-      }, embeddingServer.url);
-      await callLocalMcp(vaultDir, "relate", {
-        fromId: hubId,
-        toId: periph2Id,
-        type: "related-to",
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const summary = await callLocalMcpResponse(vaultDir, "project_memory_summary", {
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const anchors = summary.structuredContent?.["anchors"] as Array<Record<string, unknown>>;
-      const orientation = summary.structuredContent?.["orientation"] as Record<string, unknown>;
-      const primaryEntry = orientation?.["primaryEntry"] as Record<string, unknown>;
-      
-      // Verify anchor exists (hub should be anchor with 2 connections)
-      expect(anchors!.length).toBeGreaterThan(0);
-      
-      // Hub (with 2 connections) should be primaryEntry
-      expect(primaryEntry?.["id"]).toBe(hubId);
-      expect(primaryEntry?.["rationale"]).toContain("Centrality");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("falls back to the most recent project note when no anchors exist", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Older project note",
-        content: "First project note with no relationships.",
-        tags: ["integration", "fallback"],
-        summary: "Create older project note without anchors",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-
-      const recentRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Most recent project note",
-        content: "Most recent project note with no relationships.",
-        tags: ["integration", "fallback"],
-        summary: "Create most recent project note without anchors",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-      const recentId = extractRememberedId(recentRemember);
-
-      const summary = await callLocalMcpResponse(vaultDir, "project_memory_summary", {
-        cwd: repoDir,
-      }, { ollamaUrl: embeddingServer.url, disableGit: false });
-
-      const anchors = summary.structuredContent?.["anchors"] as Array<Record<string, unknown>>;
-      expect(anchors).toEqual([]);
-
-      const orientation = summary.structuredContent?.["orientation"] as Record<string, unknown>;
-      const primaryEntry = orientation?.["primaryEntry"] as Record<string, unknown>;
-      expect(primaryEntry?.["id"]).toBe(recentId);
-      expect(primaryEntry?.["title"]).toBe("Most recent project note");
-      expect(primaryEntry?.["rationale"]).toBe("Most recent note — no high-centrality anchors found");
-      // provenance is undefined when git is disabled or note has no commits; with git enabled it should be populated
-      expect(primaryEntry).toHaveProperty("confidence");
-      expect(orientation?.["suggestedNext"]).toEqual([]);
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("returns ranked suggestedNext entries with orientation rationale", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const anchorA = extractRememberedId(await callLocalMcp(vaultDir, "remember", {
-        title: "Anchor A",
-        content: "Highest-centrality anchor.",
-        tags: ["integration", "orientation", "overview"],
-        summary: "Create anchor A",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url));
-
-      const anchorB = extractRememberedId(await callLocalMcp(vaultDir, "remember", {
-        title: "Anchor B",
-        content: "Second-best anchor.",
-        tags: ["integration", "orientation", "decisions"],
-        summary: "Create anchor B",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url));
-
-      const anchorC = extractRememberedId(await callLocalMcp(vaultDir, "remember", {
-        title: "Anchor C",
-        content: "Third-best anchor.",
-        tags: ["integration", "orientation", "architecture"],
-        summary: "Create anchor C",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url));
-
-      const anchorD = extractRememberedId(await callLocalMcp(vaultDir, "remember", {
-        title: "Anchor D",
-        content: "Fourth-best anchor.",
-        tags: ["integration", "orientation", "tools"],
-        summary: "Create anchor D",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url));
-
-      const leaf1 = extractRememberedId(await callLocalMcp(vaultDir, "remember", {
-        title: "Leaf 1",
-        content: "Leaf note 1.",
-        tags: ["integration", "orientation", "decisions"],
-        summary: "Create leaf 1",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "temporary",
-      }, embeddingServer.url));
-
-      const leaf2 = extractRememberedId(await callLocalMcp(vaultDir, "remember", {
-        title: "Leaf 2",
-        content: "Leaf note 2.",
-        tags: ["integration", "orientation", "architecture"],
-        summary: "Create leaf 2",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "temporary",
-      }, embeddingServer.url));
-
-      const leaf3 = extractRememberedId(await callLocalMcp(vaultDir, "remember", {
-        title: "Leaf 3",
-        content: "Leaf note 3.",
-        tags: ["integration", "orientation", "tooling"],
-        summary: "Create leaf 3",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "temporary",
-      }, embeddingServer.url));
-
-      const leaf4 = extractRememberedId(await callLocalMcp(vaultDir, "remember", {
-        title: "Leaf 4",
-        content: "Leaf note 4.",
-        tags: ["integration", "orientation", "bugs"],
-        summary: "Create leaf 4",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "temporary",
-      }, embeddingServer.url));
-
-      await callLocalMcp(vaultDir, "relate", { fromId: anchorA, toId: leaf1, type: "related-to", cwd: repoDir }, embeddingServer.url);
-      await callLocalMcp(vaultDir, "relate", { fromId: anchorA, toId: leaf2, type: "related-to", cwd: repoDir }, embeddingServer.url);
-      await callLocalMcp(vaultDir, "relate", { fromId: anchorA, toId: leaf3, type: "related-to", cwd: repoDir }, embeddingServer.url);
-      await callLocalMcp(vaultDir, "relate", { fromId: anchorB, toId: leaf1, type: "related-to", cwd: repoDir }, embeddingServer.url);
-      await callLocalMcp(vaultDir, "relate", { fromId: anchorB, toId: leaf2, type: "related-to", cwd: repoDir }, embeddingServer.url);
-      await callLocalMcp(vaultDir, "relate", { fromId: anchorC, toId: leaf1, type: "related-to", cwd: repoDir }, embeddingServer.url);
-      await callLocalMcp(vaultDir, "relate", { fromId: anchorD, toId: leaf4, type: "related-to", cwd: repoDir }, embeddingServer.url);
-
-      const summary = await callLocalMcpResponse(vaultDir, "project_memory_summary", {
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const anchors = summary.structuredContent?.["anchors"] as Array<Record<string, unknown>>;
-      const orientation = summary.structuredContent?.["orientation"] as Record<string, unknown>;
-      const primaryEntry = orientation?.["primaryEntry"] as Record<string, unknown>;
-      const suggestedNext = orientation?.["suggestedNext"] as Array<Record<string, unknown>>;
-
-      expect(primaryEntry?.["id"]).toBe(anchorA);
-      expect(anchors.map((entry) => entry["id"])).toEqual([anchorA, ...suggestedNext.map((entry) => entry["id"])]);
-      expect(suggestedNext).toHaveLength(3);
-      expect(suggestedNext.map((entry) => entry["id"])).toEqual(anchors.slice(1, 4).map((entry) => entry["id"]));
-      expect(suggestedNext.map((entry) => entry["id"])).toEqual(expect.arrayContaining([anchorB, anchorC, anchorD]));
-      for (const entry of suggestedNext) {
-        expect(String(entry["rationale"] ?? "")).toContain("Centrality");
-        expect(String(entry["rationale"] ?? "")).toContain("connects");
-      }
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("emits taxonomy dilution warnings only when other exceeds thirty percent", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const notes = [
-        { title: "Random note 1", tags: ["integration", "warning-check"] },
-        { title: "Random note 2", tags: ["integration", "warning-check"] },
-        { title: "Decision note", tags: ["integration", "warning-check", "decisions"] },
-        { title: "Tool note", tags: ["integration", "warning-check", "tools"] },
-      ];
-
-      for (const note of notes) {
-        await callLocalMcp(vaultDir, "remember", {
-          title: note.title,
-          content: `${note.title} content.`,
-          tags: note.tags,
-          summary: `Create ${note.title}`,
-          cwd: repoDir,
-          scope: "project",
-          lifecycle: "permanent",
-        }, embeddingServer.url);
-      }
-
-      const warningSummary = await callLocalMcpResponse(vaultDir, "project_memory_summary", {
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const warningOrientation = warningSummary.structuredContent?.["orientation"] as Record<string, unknown>;
-      const warnings = warningOrientation?.["warnings"] as string[] | undefined;
-      expect(warnings).toBeDefined();
-      expect(warnings?.[0]).toContain('50% of notes in "other" bucket');
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Architecture note",
-        content: "Architecture content.",
-        tags: ["integration", "warning-check", "architecture"],
-        summary: "Create architecture note",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Quality note",
-        content: "Quality content.",
-        tags: ["integration", "warning-check", "tests"],
-        summary: "Create quality note",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Overview note",
-        content: "Overview content.",
-        tags: ["integration", "warning-check", "overview"],
-        summary: "Create overview note",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-
-      const noWarningSummary = await callLocalMcpResponse(vaultDir, "project_memory_summary", {
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const noWarningOrientation = noWarningSummary.structuredContent?.["orientation"] as Record<string, unknown>;
-      expect(noWarningOrientation?.["warnings"]).toBeUndefined();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("does not include relatedGlobal unless explicitly requested", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const hubId = extractRememberedId(await callLocalMcp(vaultDir, "remember", {
-        title: "Project hub",
-        content: "Project hub note.",
-        tags: ["integration", "related-global"],
-        summary: "Create project hub",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url));
-
-      const leafId = extractRememberedId(await callLocalMcp(vaultDir, "remember", {
-        title: "Project leaf",
-        content: "Project leaf note.",
-        tags: ["integration", "related-global", "decisions"],
-        summary: "Create project leaf",
-        cwd: repoDir,
-        scope: "project",
-        lifecycle: "permanent",
-      }, embeddingServer.url));
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Global note candidate",
-        content: "Unscoped global note that should only appear in relatedGlobal when requested.",
-        tags: ["integration", "related-global"],
-        summary: "Create related-global candidate",
-        scope: "global",
-        lifecycle: "permanent",
-      }, embeddingServer.url);
-
-      await callLocalMcp(vaultDir, "relate", {
-        fromId: hubId,
-        toId: leafId,
-        type: "related-to",
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const summary = await callLocalMcpResponse(vaultDir, "project_memory_summary", {
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      expect(summary.structuredContent?.["relatedGlobal"]).toBeUndefined();
-      expect(summary.text).not.toContain("Related Global:");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("shows only visible cross-vault relationships in memory_graph", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const privateProjectRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Graph private project memory",
-        content: "Stored in main vault but associated with the current project.",
-        tags: ["integration", "graph"],
-        summary: "Create private project memory for graph test",
-        cwd: repoDir,
-        scope: "global",
-      }, embeddingServer.url);
-      const privateProjectId = extractRememberedId(privateProjectRemember);
-
-      const sharedProjectRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Graph shared project memory",
-        content: "Stored in the project vault and linked into the graph.",
-        tags: ["integration", "graph"],
-        summary: "Create shared project memory for graph test",
-        cwd: repoDir,
-        scope: "project",
-      }, embeddingServer.url);
-      const sharedProjectId = extractRememberedId(sharedProjectRemember);
-
-      const globalRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Graph global memory",
-        content: "Unscoped global memory that should disappear from project-only graph results.",
-        tags: ["integration", "graph"],
-        summary: "Create global memory for graph test",
-        scope: "global",
-      }, embeddingServer.url);
-      const globalId = extractRememberedId(globalRemember);
-
-      await callLocalMcp(vaultDir, "relate", {
-        fromId: privateProjectId,
-        toId: sharedProjectId,
-        type: "related-to",
-        bidirectional: true,
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      await callLocalMcp(vaultDir, "relate", {
-        fromId: privateProjectId,
-        toId: globalId,
-        type: "explains",
-        bidirectional: true,
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      const graphAll = await callLocalMcpResponse(vaultDir, "memory_graph", {
-        cwd: repoDir,
-        scope: "all",
-        storedIn: "any",
-        limit: 10,
-      }, embeddingServer.url);
-
-      expect(graphAll.text).toContain(privateProjectId);
-      expect(graphAll.text).toContain(sharedProjectId);
-      expect(graphAll.text).toContain(globalId);
-      const allNodes = graphAll.structuredContent?.["nodes"] as Array<Record<string, unknown>>;
-      const privateNode = allNodes.find((node) => node["id"] === privateProjectId);
-      expect(privateNode).toBeTruthy();
-      expect((privateNode?.["edges"] as Array<Record<string, unknown>>).map((edge) => edge["toId"]).sort()).toEqual([
-        globalId,
-        sharedProjectId,
-      ].sort());
-
-      const graphProject = await callLocalMcpResponse(vaultDir, "memory_graph", {
-        cwd: repoDir,
-        scope: "project",
-        storedIn: "any",
-        limit: 10,
-      }, embeddingServer.url);
-
-      expect(graphProject.text).toContain(privateProjectId);
-      expect(graphProject.text).toContain(sharedProjectId);
-      expect(graphProject.text).not.toContain(globalId);
-      const projectNodes = graphProject.structuredContent?.["nodes"] as Array<Record<string, unknown>>;
-      const projectPrivateNode = projectNodes.find((node) => node["id"] === privateProjectId);
-      expect((projectPrivateNode?.["edges"] as Array<Record<string, unknown>>).map((edge) => edge["toId"])).toEqual([
-        sharedProjectId,
-      ]);
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("updates an existing memory through the MCP and persists the edited content", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Initial integration note",
-        content: "Original content for update flow.",
-        tags: ["integration", "original"],
-        summary: "Create note for MCP update test",
-        scope: "global",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-      const updateText = await callLocalMcp(vaultDir, "update", {
-        id: noteId,
-        title: "Updated integration note",
-        content: "Updated content for update flow.",
-        tags: ["integration", "updated"],
-        summary: "Verify MCP update persists content changes",
-      }, embeddingServer.url);
-
-      expect(updateText).toContain(`Updated memory '${noteId}'`);
-
-      const notePath = path.join(vaultDir, "notes", `${noteId}.md`);
-      const embeddingPath = path.join(vaultDir, "embeddings", `${noteId}.json`);
-      const noteContents = await readFile(notePath, "utf-8");
-
-      expect(noteContents).toContain("title: Updated integration note");
-      expect(noteContents).toContain("Updated content for update flow.");
-      expect(noteContents).toContain("- integration");
-      expect(noteContents).toContain("- updated");
-      await expect(stat(embeddingPath)).resolves.toBeDefined();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("preserves lifecycle on update unless explicitly changed", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Temporary integration lifecycle note",
-        content: "Initial temporary plan note.",
-        tags: ["integration", "plan"],
-        lifecycle: "temporary",
-        summary: "Create temporary note for lifecycle update test",
-        scope: "global",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-      const notePath = path.join(vaultDir, "notes", `${noteId}.md`);
-
-      let noteContents = await readFile(notePath, "utf-8");
-      expect(noteContents).toContain("lifecycle: temporary");
-
-      await callLocalMcp(vaultDir, "update", {
-        id: noteId,
-        content: "Still temporary after a regular update.",
-        summary: "Verify lifecycle is preserved when omitted",
-      }, embeddingServer.url);
-
-      noteContents = await readFile(notePath, "utf-8");
-      expect(noteContents).toContain("lifecycle: temporary");
-
-      await callLocalMcp(vaultDir, "update", {
-        id: noteId,
-        lifecycle: "permanent",
-        summary: "Promote lifecycle to permanent explicitly",
-      }, embeddingServer.url);
-
-      noteContents = await readFile(notePath, "utf-8");
-      expect(noteContents).toContain("lifecycle: permanent");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("cleans related notes when forgetting a linked memory", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const firstRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "First linked note",
-        content: "First note in relation test.",
-        tags: ["integration"],
-        summary: "Create first note for relate and forget test",
-        scope: "global",
-      }, embeddingServer.url);
-      const secondRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Second linked note",
-        content: "Second note in relation test.",
-        tags: ["integration"],
-        summary: "Create second note for relate and forget test",
-        scope: "global",
-      }, embeddingServer.url);
-
-      const firstId = extractRememberedId(firstRemember);
-      const secondId = extractRememberedId(secondRemember);
-
-      const relateText = await callLocalMcp(vaultDir, "relate", {
-        fromId: firstId,
-        toId: secondId,
-        type: "related-to",
-      }, embeddingServer.url);
-      expect(relateText).toContain(`Linked \`${firstId}\` ↔ \`${secondId}\` (related-to)`);
-
-      const beforeForget = await readFile(path.join(vaultDir, "notes", `${secondId}.md`), "utf-8");
-      expect(beforeForget).toContain(firstId);
-
-      const forgetText = await callLocalMcp(vaultDir, "forget", { id: firstId }, embeddingServer.url);
-      expect(forgetText).toContain(`Forgotten '${firstId}'`);
-
-      await expect(stat(path.join(vaultDir, "notes", `${firstId}.md`))).rejects.toThrow();
-      const survivor = await readFile(path.join(vaultDir, "notes", `${secondId}.md`), "utf-8");
-      expect(survivor).not.toContain(firstId);
-      expect(survivor).toContain("Second linked note");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("removes bidirectional cross-vault relationships via unrelate", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const mainRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Cross vault main note",
-        content: "Stored privately in main but tied to the current project.",
-        tags: ["integration", "relations"],
-        summary: "Create main-vault note for unrelate test",
-        cwd: repoDir,
-        scope: "global",
-      }, embeddingServer.url);
-      const mainId = extractRememberedId(mainRemember);
-
-      const projectRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Cross vault project note",
-        content: "Stored in the project vault and linked to the main-vault note.",
-        tags: ["integration", "relations"],
-        summary: "Create project-vault note for unrelate test",
-        cwd: repoDir,
-        scope: "project",
-      }, embeddingServer.url);
-      const projectId = extractRememberedId(projectRemember);
-
-      const relateText = await callLocalMcp(vaultDir, "relate", {
-        fromId: mainId,
-        toId: projectId,
-        type: "related-to",
-        bidirectional: true,
-        cwd: repoDir,
-      }, embeddingServer.url);
-      expect(relateText).toContain(`Linked \`${mainId}\` ↔ \`${projectId}\``);
-
-      const unrelated = await callLocalMcpResponse(vaultDir, "unrelate", {
-        fromId: mainId,
-        toId: projectId,
-        bidirectional: true,
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      expect(unrelated.text).toContain(`Removed relationship between \`${mainId}\` and \`${projectId}\``);
-      const modified = unrelated.structuredContent?.["notesModified"] as string[];
-      expect(modified.sort()).toEqual([mainId, projectId].sort());
-
-      const mainContents = await readFile(path.join(vaultDir, "notes", `${mainId}.md`), "utf-8");
-      const projectContents = await readFile(path.join(repoDir, ".mnemonic", "notes", `${projectId}.md`), "utf-8");
-      expect(mainContents).not.toContain(projectId);
-      expect(projectContents).not.toContain(mainId);
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("deletes temporary source notes and creates a permanent target on consolidation", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const firstRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Temporary plan A",
-        content: "Temporary implementation plan A.",
-        tags: ["integration", "plan"],
-        lifecycle: "temporary",
-        summary: "Create first temporary plan note",
-        cwd: repoDir,
-        scope: "project",
-      }, embeddingServer.url);
-      const secondRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Temporary plan B",
-        content: "Temporary implementation plan B.",
-        tags: ["integration", "plan"],
-        lifecycle: "temporary",
-        summary: "Create second temporary plan note",
-        cwd: repoDir,
-        scope: "project",
-      }, embeddingServer.url);
-
-      const firstId = extractRememberedId(firstRemember);
-      const secondId = extractRememberedId(secondRemember);
-
-      const consolidateText = await callLocalMcp(vaultDir, "consolidate", {
-        cwd: repoDir,
-        strategy: "execute-merge",
-        mergePlan: {
-          sourceIds: [firstId, secondId],
-          targetTitle: "Consolidated implementation plan",
-        },
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-
-      expect(consolidateText).toContain("Mode: delete");
-      expect(consolidateText).toContain("Source notes deleted.");
-
-      await expect(stat(path.join(repoDir, ".mnemonic", "notes", `${firstId}.md`))).rejects.toThrow();
-      await expect(stat(path.join(repoDir, ".mnemonic", "notes", `${secondId}.md`))).rejects.toThrow();
-
-      const consolidatedIdMatch = consolidateText.match(/Consolidated \d+ notes into '([^']+)'/);
-      expect(consolidatedIdMatch).toBeTruthy();
-      const consolidatedId = consolidatedIdMatch![1]!;
-      const consolidatedPath = path.join(repoDir, ".mnemonic", "notes", `${consolidatedId}.md`);
-      const consolidatedContents = await readFile(consolidatedPath, "utf-8");
-      expect(consolidatedContents).toContain("lifecycle: permanent");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("uses custom content body when mergePlan.content is provided", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const planRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Implementation plan",
-        content: "Plan: do X, Y, Z.",
-        tags: ["plan"],
-        lifecycle: "temporary",
-        summary: "Create temporary plan note",
-        cwd: repoDir,
-        scope: "project",
-      }, embeddingServer.url);
-      const outcomeRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Implementation outcome",
-        content: "Outcome: X, Y, Z done. Changed A and B.",
-        tags: ["outcome"],
-        lifecycle: "temporary",
-        summary: "Create temporary outcome note",
-        cwd: repoDir,
-        scope: "project",
-      }, embeddingServer.url);
-
-      const planId = extractRememberedId(planRemember);
-      const outcomeId = extractRememberedId(outcomeRemember);
-
-      const distilledContent = "Implemented persistence-status reporting. Changed A and B. Pattern: return structured status from all mutating tools.";
-      const consolidateText = await callLocalMcp(vaultDir, "consolidate", {
-        cwd: repoDir,
-        strategy: "execute-merge",
-        mergePlan: {
-          sourceIds: [planId, outcomeId],
-          targetTitle: "Persistence status reporting",
-          content: distilledContent,
-        },
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-
-      expect(consolidateText).toContain("Mode: delete");
-
-      const consolidatedIdMatch = consolidateText.match(/Consolidated \d+ notes into '([^']+)'/);
-      expect(consolidatedIdMatch).toBeTruthy();
-      const consolidatedId = consolidatedIdMatch![1]!;
-      const consolidatedPath = path.join(repoDir, ".mnemonic", "notes", `${consolidatedId}.md`);
-      const consolidatedContents = await readFile(consolidatedPath, "utf-8");
-
-      expect(consolidatedContents).toContain(distilledContent);
-      expect(consolidatedContents).not.toContain("## Consolidated from:");
-      expect(consolidatedContents).not.toContain("Plan: do X, Y, Z.");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("reuses an existing execute-merge target on retry instead of creating a duplicate", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const firstRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Retry source A",
-        content: "First source note for execute-merge retry handling.",
-        tags: ["integration", "merge"],
-        lifecycle: "permanent",
-        summary: "Create first source for execute-merge idempotency test",
-        cwd: repoDir,
-        scope: "project",
-      }, embeddingServer.url);
-      const secondRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Retry source B",
-        content: "Second source note for execute-merge retry handling.",
-        tags: ["integration", "merge"],
-        lifecycle: "permanent",
-        summary: "Create second source for execute-merge idempotency test",
-        cwd: repoDir,
-        scope: "project",
-      }, embeddingServer.url);
-
-      const firstId = extractRememberedId(firstRemember);
-      const secondId = extractRememberedId(secondRemember);
-
-      const firstMerge = await callLocalMcp(vaultDir, "consolidate", {
-        cwd: repoDir,
-        strategy: "execute-merge",
-        mode: "supersedes",
-        mergePlan: {
-          sourceIds: [firstId, secondId],
-          targetTitle: "Retry-safe consolidated note",
-          content: "First consolidated body.",
-        },
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-
-      const firstTargetIdMatch = firstMerge.match(/Consolidated \d+ notes into '([^']+)'/);
-      expect(firstTargetIdMatch).toBeTruthy();
-      const firstTargetId = firstTargetIdMatch![1]!;
-
-      const secondMerge = await callLocalMcp(vaultDir, "consolidate", {
-        cwd: repoDir,
-        strategy: "execute-merge",
-        mode: "supersedes",
-        mergePlan: {
-          sourceIds: [firstId, secondId],
-          targetTitle: "Retry-safe consolidated note",
-          content: "Updated consolidated body after retry.",
-        },
-        allowProtectedBranch: true,
-      }, embeddingServer.url);
-
-      const secondTargetIdMatch = secondMerge.match(/Consolidated \d+ notes into '([^']+)'/);
-      expect(secondTargetIdMatch).toBeTruthy();
-      const secondTargetId = secondTargetIdMatch![1]!;
-
-      expect(secondTargetId).toBe(firstTargetId);
-      expect(secondMerge).toContain("Idempotency: reused existing target note.");
-
-      const targetPath = path.join(repoDir, ".mnemonic", "notes", `${firstTargetId}.md`);
-      const targetContents = await readFile(targetPath, "utf-8");
-      expect(targetContents).toContain("Updated consolidated body after retry.");
-      expect(targetContents).not.toContain("First consolidated body.");
-
-      const noteFiles = await readdir(path.join(repoDir, ".mnemonic", "notes"));
-      const matchingTargets = noteFiles.filter((file) => file.startsWith("retry-safe-consolidated-note-") && file.endsWith(".md"));
-      expect(matchingTargets).toHaveLength(1);
-
-      const firstSourceContents = await readFile(path.join(repoDir, ".mnemonic", "notes", `${firstId}.md`), "utf-8");
-      const secondSourceContents = await readFile(path.join(repoDir, ".mnemonic", "notes", `${secondId}.md`), "utf-8");
-      expect(firstSourceContents.match(new RegExp(firstTargetId, "g")) ?? []).toHaveLength(1);
-      expect(secondSourceContents.match(new RegExp(firstTargetId, "g")) ?? []).toHaveLength(1);
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("reports sync status cleanly when git syncing is disabled", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Sync no remote note",
-        content: "Created in main vault with embedding unavailable so sync can backfill it.",
-        scope: "global",
-        summary: "Seed main-vault note without embedding for sync test",
-      }, { ollamaUrl: "http://127.0.0.1:9" });
-
-      const noteId = extractRememberedId(rememberText);
-      await expect(stat(path.join(vaultDir, "embeddings", `${noteId}.json`))).rejects.toThrow();
-
-      const syncText = await callLocalMcp(vaultDir, "sync", { cwd: repoDir }, embeddingServer.url);
-
-      expect(syncText).toContain("main vault: no remote configured — git sync skipped.");
-      expect(syncText).toContain("main vault: embedded 1 note(s) (including any missing local embeddings).");
-      expect(syncText).toContain("project vault: no .mnemonic/ found — skipped.");
-      await expect(stat(path.join(vaultDir, "embeddings", `${noteId}.json`))).resolves.toBeDefined();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("backfills missing project embeddings during sync on a fresh clone", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const remoteDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-remote-"));
-    const sourceDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-source-"));
-    const cloneDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-clone-"));
-    tempDirs.push(vaultDir, remoteDir, sourceDir, cloneDir);
-
-    await execFileAsync("git", ["init", "--bare", remoteDir]);
-    await execFileAsync("git", ["init", "-b", "main"], { cwd: sourceDir });
-    await execFileAsync("git", ["remote", "add", "origin", remoteDir], { cwd: sourceDir });
-
-    await mkdir(path.join(sourceDir, ".mnemonic", "notes"), { recursive: true });
-    await writeFile(path.join(sourceDir, ".mnemonic", ".gitignore"), "embeddings/\n", "utf-8");
-    await writeFile(
-      path.join(sourceDir, ".mnemonic", "notes", "fresh-clone-note.md"),
-      `---\ntitle: Fresh clone note\ntags: []\nlifecycle: permanent\ncreatedAt: 2026-03-09T00:00:00.000Z\nupdatedAt: 2026-03-09T00:00:00.000Z\nmemoryVersion: 1\n---\n\nThis note should be embedded during sync on a fresh machine.`,
-      "utf-8",
-    );
-
-    await execFileAsync("git", ["add", "."], { cwd: sourceDir });
-    await execFileAsync(
-      "git",
-      ["-c", "user.name=Test User", "-c", "user.email=test@example.com", "commit", "-m", "seed project mnemonic notes"],
-      { cwd: sourceDir },
-    );
-    await execFileAsync("git", ["push", "-u", "origin", "main"], { cwd: sourceDir });
-
-    await execFileAsync("git", ["clone", "--branch", "main", remoteDir, cloneDir]);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const beforeEmbedding = path.join(cloneDir, ".mnemonic", "embeddings", "fresh-clone-note.json");
-      await expect(stat(beforeEmbedding)).rejects.toThrow();
-
-      const syncText = await callLocalMcp(
-        vaultDir,
-        "sync",
-        { cwd: cloneDir },
-        { ollamaUrl: embeddingServer.url, disableGit: false },
-      );
-
-      expect(syncText).toContain("project vault: ↓ no new notes from remote.");
-      expect(syncText).toContain("project vault: embedded 1 note(s) (including any missing local embeddings).");
-      await expect(stat(beforeEmbedding)).resolves.toBeDefined();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 20000);
-
-  it("returns structured migration metadata from list_migrations", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-
-    const response = await callLocalMcpResponse(vaultDir, "list_migrations", {});
-
-    expect(response.text).toContain("Available migrations:");
-
-    const structured = response.structuredContent;
-    expect(structured?.["action"]).toBe("migration_list");
-    expect(structured?.["totalPending"]).toBeTypeOf("number");
-
-    const available = structured?.["available"] as Array<Record<string, unknown>>;
-    expect(Array.isArray(available)).toBe(true);
-    expect(available.length).toBeGreaterThan(0);
-    expect(available.some((migration) => migration["name"] === "v0.1.0-backfill-memory-versions")).toBe(true);
-
-    const vaults = structured?.["vaults"] as Array<Record<string, unknown>>;
-    expect(Array.isArray(vaults)).toBe(true);
-    expect(vaults.length).toBeGreaterThan(0);
-  }, 15000);
-
-  it("keeps migration and graph structured outputs aligned with their schemas", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Schema audit graph note",
-        content: "Used to validate structured MCP output schemas.",
-        tags: ["integration"],
-        scope: "global",
-        summary: "Create note for structured output schema audit",
-      }, embeddingServer.url);
-
-      extractRememberedId(rememberText);
-
-      const migrationList = await callLocalMcpResponse(vaultDir, "list_migrations", {});
-      expect(() => MigrationListResultSchema.parse(migrationList.structuredContent)).not.toThrow();
-
-      const executeMigration = await callLocalMcpResponse(vaultDir, "execute_migration", {
-        migrationName: "v0.1.0-backfill-memory-versions",
-        dryRun: true,
-        backup: true,
-      });
-      expect(() => MigrationExecuteResultSchema.parse(executeMigration.structuredContent)).not.toThrow();
-
-      const memoryGraph = await callLocalMcpResponse(vaultDir, "memory_graph", {});
-      const graph = MemoryGraphResultSchema.parse(memoryGraph.structuredContent);
-      expect(Array.isArray(graph.nodes)).toBe(true);
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("validates consolidate find-clusters structured output includes themeGroups and relationshipClusters", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const remember1 = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Consolidate cluster note A",
-        content: "First note for cluster analysis testing with similar content.",
-        tags: ["cluster-test"],
-        scope: "global",
-        summary: "Create first note for consolidate cluster test",
-      }, embeddingServer.url);
-      const remember2 = await callLocalMcpResponse(vaultDir, "remember", {
-        title: "Consolidate cluster note B",
-        content: "Second note for cluster analysis testing with similar content.",
-        tags: ["cluster-test"],
-        scope: "global",
-        summary: "Create second note for consolidate cluster test",
-      }, embeddingServer.url);
-
-      extractRememberedId(remember1.text);
-      extractRememberedId(remember2.text);
-
-      const consolidateResult = await callLocalMcpResponse(vaultDir, "consolidate", {
-        strategy: "find-clusters",
-      });
-
-      const result = ConsolidateResultSchema.parse(consolidateResult.structuredContent);
-
-      expect(result.strategy).toBe("find-clusters");
-      expect(Array.isArray(result.themeGroups)).toBe(true);
-      expect(Array.isArray(result.relationshipClusters)).toBe(true);
-
-      expect(result.themeGroups?.length).toBeGreaterThan(0);
-      if (result.themeGroups && result.themeGroups.length > 0) {
-        const group = result.themeGroups[0]!;
-        expect(typeof group.name).toBe("string");
-        expect(typeof group.count).toBe("number");
-        expect(Array.isArray(group.examples)).toBe(true);
-      }
-
-      expect(result.relationshipClusters?.length).toBeGreaterThanOrEqual(0);
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("fetches full note content via get by exact id", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Get tool test note",
-        content: "Content only accessible via get.",
-        tags: ["get-test"],
-        scope: "global",
-        summary: "Seed note for get tool test",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-
-      const response = await callLocalMcpResponse(vaultDir, "get", {
-        ids: [noteId],
-      }, embeddingServer.url);
-
-      expect(response.text).toContain("Get tool test note");
-      expect(response.text).toContain("Content only accessible via get.");
-
-      const structured = response.structuredContent;
-      expect(structured?.["action"]).toBe("got");
-      expect(structured?.["count"]).toBe(1);
-      const notes = structured?.["notes"] as Array<Record<string, unknown>>;
-      expect(notes).toHaveLength(1);
-      expect(notes[0]?.["id"]).toBe(noteId);
-      expect(notes[0]?.["title"]).toBe("Get tool test note");
-      expect(notes[0]?.["content"]).toContain("Content only accessible via get.");
-      expect(notes[0]?.["vault"]).toBe("main-vault");
-      expect(structured?.["notFound"]).toEqual([]);
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("reports not found ids from get", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-
-    const response = await callLocalMcpResponse(vaultDir, "get", {
-      ids: ["nonexistent-id-abc123"],
-    });
-
-    const structured = response.structuredContent;
-    expect(structured?.["action"]).toBe("got");
-    expect(structured?.["count"]).toBe(0);
-    expect(structured?.["notFound"]).toEqual(["nonexistent-id-abc123"]);
-  }, 15000);
-
-  it("locates a memory via where_is_memory", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Where is memory test",
-        content: "A note to locate.",
-        scope: "global",
-        summary: "Seed note for where_is_memory test",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-
-      const response = await callLocalMcpResponse(vaultDir, "where_is_memory", {
-        id: noteId,
-      }, embeddingServer.url);
-
-      expect(response.text).toContain("Where is memory test");
-      expect(response.text).toContain("main-vault");
-
-      const structured = response.structuredContent;
-      expect(structured?.["action"]).toBe("located");
-      expect(structured?.["id"]).toBe(noteId);
-      expect(structured?.["title"]).toBe("Where is memory test");
-      expect(structured?.["vault"]).toBe("main-vault");
-      expect(structured?.["relatedCount"]).toBe(0);
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("recall backfills a missing embedding and returns the note", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-
-    await mkdir(path.join(vaultDir, "notes"), { recursive: true });
-    await writeFile(
-      path.join(vaultDir, "notes", "backfill-recall-note.md"),
-      `---\ntitle: Lazy backfill recall note\ntags: [integration]\nlifecycle: permanent\ncreatedAt: 2026-01-01T00:00:00.000Z\nupdatedAt: 2026-01-01T00:00:00.000Z\nmemoryVersion: 1\n---\n\nThis note has no embedding yet and should be found via recall.`,
-      "utf-8",
-    );
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const recallText = await callLocalMcp(vaultDir, "recall", {
-        query: "lazy backfill recall",
-      }, embeddingServer.url);
-
-      expect(recallText).toContain("Lazy backfill recall note");
-      await expect(stat(path.join(vaultDir, "embeddings", "backfill-recall-note.json"))).resolves.toBeDefined();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("recall re-embeds a stale note edited after its embedding was written", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Staleness detection note",
-        content: "Original content before direct edit.",
-        tags: ["integration"],
-        scope: "global",
-        summary: "Seed note for staleness detection test",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-      const embeddingPath = path.join(vaultDir, "embeddings", `${noteId}.json`);
-
-      // Back-date the embedding so it appears stale
-      const embeddingRaw = await readFile(embeddingPath, "utf-8");
-      const embeddingJson = JSON.parse(embeddingRaw) as Record<string, unknown>;
-      embeddingJson["updatedAt"] = "2020-01-01T00:00:00.000Z";
-      await writeFile(embeddingPath, JSON.stringify(embeddingJson), "utf-8");
-
-      // recall should detect stale embedding and regenerate it
-      await callLocalMcp(vaultDir, "recall", { query: "staleness detection" }, embeddingServer.url);
-
-      const afterRaw = await readFile(embeddingPath, "utf-8");
-      const afterJson = JSON.parse(afterRaw) as Record<string, unknown>;
-      expect(afterJson["updatedAt"]).not.toBe("2020-01-01T00:00:00.000Z");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("recall returns existing results when Ollama is down (backfill fails silently)", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Offline recall note",
-        content: "This note has an embedding and should be found even when Ollama is down.",
-        tags: ["integration"],
-        scope: "global",
-        summary: "Seed note for offline recall test",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-
-      // recall with Ollama down — embed(query) will fail, so the whole call fails
-      // But if we have a note without an embedding, the backfill should fail silently
-      // and recall should still return existing notes
-      // Create a second note without an embedding
-      await mkdir(path.join(vaultDir, "notes"), { recursive: true });
-      await writeFile(
-        path.join(vaultDir, "notes", "no-embedding-note.md"),
-        `---\ntitle: Note without embedding\ntags: [integration]\nlifecycle: permanent\ncreatedAt: 2026-01-01T00:00:00.000Z\nupdatedAt: 2026-01-01T00:00:00.000Z\nmemoryVersion: 1\n---\n\nThis note has no embedding.`,
-        "utf-8",
-      );
-
-      // recall with working Ollama: backfill for the no-embedding note should succeed, existing note also returned
-      const recallText = await callLocalMcp(vaultDir, "recall", {
-        query: "offline recall note",
-      }, embeddingServer.url);
-
-      expect(recallText).toContain("Offline recall note");
-      // The no-embedding note got backfilled too
-      await expect(stat(path.join(vaultDir, "embeddings", "no-embedding-note.json"))).resolves.toBeDefined();
-      // Original note embedding still present
-      await expect(stat(path.join(vaultDir, "embeddings", `${noteId}.json`))).resolves.toBeDefined();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("suggests note-oriented canonical tags by default", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      // Create notes with different tags and lifecycles
-      for (let i = 0; i < 4; i++) {
-        await callLocalMcp(vaultDir, "remember", {
-          title: `Broad design note ${i + 1}`,
-          content: "Design guidance for discover tags workflow and architecture decisions.",
-          tags: ["design", "architecture"],
-          lifecycle: "permanent",
-          scope: "global",
-          summary: "Create broad design note for ranking pressure",
-        }, embeddingServer.url);
-      }
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Tag discovery test note one",
-        content: "First note for tag discovery with permanent lifecycle.",
-        tags: ["test-tag", "permanent-only", "discovery", "discover_tags"],
-        lifecycle: "permanent",
-        scope: "global",
-        summary: "Create first note for tag discovery test",
-      }, embeddingServer.url);
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Tag discovery test note two",
-        content: "Second note for tag discovery also using test-tag.",
-        tags: ["test-tag", "mixed-lifecycle", "discovery", "discover_tags"],
-        lifecycle: "permanent",
-        scope: "global",
-        summary: "Create second note for tag discovery test",
-      }, embeddingServer.url);
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Temporary tag discovery note",
-        content: "Temporary note with temp-only tag.",
-        tags: ["temp-only", "discovery"],
-        lifecycle: "temporary",
-        scope: "global",
-        summary: "Create temporary note for tag discovery test",
-      }, embeddingServer.url);
-
-      const response = await callLocalMcpResponse(vaultDir, "discover_tags", {
-        title: "Tune discover_tags ranking",
-        content: "Make discover_tags suggestions prefer specific tags over generic design tags.",
-        scope: "global",
-      }, embeddingServer.url);
-
-      const structured = response.structuredContent;
-      expect(structured?.["action"]).toBe("tags_discovered");
-      expect(structured?.["mode"]).toBe("suggest");
-      expect(structured?.["totalTags"]).toBeGreaterThan(0);
-      expect(structured?.["totalNotes"]).toBe(7);
-
-      const tags = structured?.["recommendedTags"] as Array<Record<string, unknown>>;
-      expect(Array.isArray(tags)).toBe(true);
-      expect(tags.length).toBeGreaterThan(0);
-      expect(tags.length).toBeLessThanOrEqual(10);
-      expect(structured?.["tags"]).toBeUndefined();
-      const rankedNames = tags.map((t) => String(t["tag"]));
-
-      // Find the test-tag (should have usageCount 2)
-      const testTag = tags.find((t) => t["tag"] === "test-tag");
-      expect(testTag).toBeDefined();
-      expect(testTag?.["usageCount"]).toBe(2);
-      expect(typeof testTag?.["example"]).toBe("string");
-      expect(testTag?.["isTemporaryOnly"]).toBe(false);
-      expect((testTag?.["lifecycleTypes"] as string[])).toContain("permanent");
-
-      // Find the temp-only tag (should have isTemporaryOnly true)
-      const tempTag = tags.find((t) => t["tag"] === "temp-only");
-      expect(tempTag).toBeDefined();
-      expect(tempTag?.["isTemporaryOnly"]).toBe(true);
-      expect((tempTag?.["lifecycleTypes"] as string[])).toContain("temporary");
-      expect((tempTag?.["lifecycleTypes"] as string[])).not.toContain("permanent");
-
-      // Find the discovery tag (should have mixed lifecycles)
-      const discoveryTag = tags.find((t) => t["tag"] === "discovery");
-      expect(discoveryTag).toBeDefined();
-      expect(discoveryTag?.["usageCount"]).toBe(3);
-      expect(discoveryTag?.["isTemporaryOnly"]).toBe(false);
-      expect((discoveryTag?.["lifecycleTypes"] as string[])).toContain("permanent");
-      expect((discoveryTag?.["lifecycleTypes"] as string[])).toContain("temporary");
-      expect(rankedNames.indexOf("test-tag")).toBeGreaterThanOrEqual(0);
-      expect(rankedNames.indexOf("discover_tags")).toBeGreaterThanOrEqual(0);
-      if (rankedNames.includes("design")) {
-        expect(rankedNames.indexOf("discover_tags")).toBeLessThan(rankedNames.indexOf("design"));
-      }
-      if (rankedNames.includes("architecture")) {
-        expect(rankedNames.indexOf("discover_tags")).toBeLessThan(rankedNames.indexOf("architecture"));
-      }
-
-      // Verify response text contains useful information
-      expect(response.text).toContain("Suggested tags");
-      expect(response.text).toContain("test-tag");
-      expect(response.text).toContain("discovery");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("supports explicit browse mode for broader tag inventory output", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Browse mode note one",
-        content: "Contains browse-tag and discovery.",
-        tags: ["browse-tag", "discovery"],
-        lifecycle: "permanent",
-        scope: "global",
-        summary: "Create first browse mode note",
-      }, embeddingServer.url);
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Browse mode note two",
-        content: "Contains browse-tag and temp-only.",
-        tags: ["browse-tag", "temp-only"],
-        lifecycle: "temporary",
-        scope: "global",
-        summary: "Create second browse mode note",
-      }, embeddingServer.url);
-
-      const response = await callLocalMcpResponse(vaultDir, "discover_tags", {
-        mode: "browse",
-        scope: "global",
-      }, embeddingServer.url);
-
-      const structured = response.structuredContent;
-      expect(structured?.["action"]).toBe("tags_discovered");
-      expect(structured?.["mode"]).toBe("browse");
-      const tags = structured?.["tags"] as Array<Record<string, unknown>>;
-      expect(Array.isArray(tags)).toBe(true);
-      expect(tags.find((t) => t["tag"] === "browse-tag")).toBeDefined();
-      expect(tags.find((t) => t["tag"] === "temp-only")?.["isTemporaryOnly"]).toBe(true);
-      expect(structured?.["recommendedTags"]).toBeUndefined();
-      expect(response.text).toContain("Tags (scope: global)");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("keeps broad canonical tags near the top when the prompt is generic", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Architecture decision record",
-        content: "Design note about project structure and architecture guidance.",
-        tags: ["design", "architecture"],
-        lifecycle: "permanent",
-        scope: "global",
-        summary: "Create broad design note",
-      }, embeddingServer.url);
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Project workflow note",
-        content: "Workflow guidance for organizing project knowledge.",
-        tags: ["workflow", "design"],
-        lifecycle: "permanent",
-        scope: "global",
-        summary: "Create workflow note",
-      }, embeddingServer.url);
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Temporary scratchpad",
-        content: "Temporary working note with short-lived cleanup context.",
-        tags: ["temporary-notes", "implemented"],
-        lifecycle: "temporary",
-        scope: "global",
-        summary: "Create temporary noise note",
-      }, embeddingServer.url);
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Implementation status",
-        content: "A consolidated implemented note about finished work items.",
-        tags: ["implemented", "consolidated"],
-        lifecycle: "permanent",
-        scope: "global",
-        summary: "Create implementation status noise note",
-      }, embeddingServer.url);
-
-      const response = await callLocalMcpResponse(vaultDir, "discover_tags", {
-        title: "Improve project notes",
-        content: "Make project knowledge easier to use across sessions.",
-        scope: "global",
-      }, embeddingServer.url);
-
-      const structured = response.structuredContent;
-      expect(structured?.["mode"]).toBe("suggest");
-      const rankedTags = structured?.["recommendedTags"] as Array<Record<string, unknown>>;
-      expect(Array.isArray(rankedTags)).toBe(true);
-      const rankedNames = rankedTags.map((tag) => String(tag["tag"]));
-      expect(rankedNames.length).toBeGreaterThan(0);
-      expect(["design", "architecture", "workflow"]).toContain(rankedNames[0]);
-      if (rankedNames.includes("temporary-notes")) {
-        expect(rankedNames.indexOf("design")).toBeLessThan(rankedNames.indexOf("temporary-notes"));
-      }
-      if (rankedNames.includes("implemented")) {
-        expect(rankedNames.indexOf("design")).toBeLessThan(rankedNames.indexOf("implemented"));
-      }
-      if (rankedNames.includes("consolidated")) {
-        expect(rankedNames.indexOf("design")).toBeLessThan(rankedNames.indexOf("consolidated"));
-      }
-      expect(structured?.["tags"]).toBeUndefined();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("prefers an exact specific tag even when it only exists once", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      for (let i = 0; i < 4; i++) {
-        await callLocalMcp(vaultDir, "remember", {
-          title: `Broad design note ${i + 1}`,
-          content: "General design and architecture guidance.",
-          tags: ["design", "architecture"],
-          lifecycle: "permanent",
-          scope: "global",
-          summary: "Create broad design baseline note",
-        }, embeddingServer.url);
-      }
-
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Cache invalidation issue",
-        content: "A focused note about cache invalidation behavior.",
-        tags: ["cache-invalidation"],
-        lifecycle: "permanent",
-        scope: "global",
-        summary: "Create one-off specific tag note",
-      }, embeddingServer.url);
-
-      const response = await callLocalMcpResponse(vaultDir, "discover_tags", {
-        title: "Fix cache-invalidation behavior",
-        content: "Investigate cache-invalidation for stale project data.",
-        scope: "global",
-      }, embeddingServer.url);
-
-      const structured = response.structuredContent;
-      expect(structured?.["mode"]).toBe("suggest");
-      const rankedTags = structured?.["recommendedTags"] as Array<Record<string, unknown>>;
-      expect(Array.isArray(rankedTags)).toBe(true);
-      const rankedNames = rankedTags.map((tag) => String(tag["tag"]));
-      expect(rankedNames.indexOf("cache-invalidation")).toBeGreaterThanOrEqual(0);
-      if (rankedNames.includes("design")) {
-        expect(rankedNames.indexOf("cache-invalidation")).toBeLessThan(rankedNames.indexOf("design"));
-      }
-      if (rankedNames.includes("architecture")) {
-        expect(rankedNames.indexOf("cache-invalidation")).toBeLessThan(rankedNames.indexOf("architecture"));
-      }
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("merges a global note and a project-associated note in a single execute-merge call", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      // A purely global note — no project association, no cwd
-      const globalRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Cross scope source A (global)",
-        content: "A purely global note with no project association.",
-        tags: ["integration", "cross-scope"],
-        lifecycle: "permanent",
-        summary: "Create global note for cross-scope consolidation test",
-        scope: "global",
-      }, embeddingServer.url);
-      const globalId = extractRememberedId(globalRemember);
-
-      // A project-associated note stored in main-vault (private/global scope with cwd)
-      const projectRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Cross scope source B (project-associated, main-vault)",
-        content: "A project-associated note stored privately in main-vault.",
-        tags: ["integration", "cross-scope"],
-        lifecycle: "permanent",
-        summary: "Create project-associated note for cross-scope consolidation test",
-        cwd: repoDir,
-        scope: "global",
-      }, embeddingServer.url);
-      const projectAssociatedId = extractRememberedId(projectRemember);
-
-      // Both notes are in main-vault but have different project associations.
-      // A single execute-merge call must resolve both — previously only one scope
-      // was searched and the other source was reported as not found.
-      // Consolidate without cwd — no project context, but both notes live in main-vault.
-      // Previously this failed because the global-scope filter excluded the project-associated note.
-      const consolidateText = await callLocalMcp(vaultDir, "consolidate", {
-        strategy: "execute-merge",
-        mode: "delete",
-        mergePlan: {
-          sourceIds: [globalId, projectAssociatedId],
-          targetTitle: "Cross scope consolidated note",
-          content: "Merged content from a global note and a project-associated note.",
-        },
-      }, embeddingServer.url);
-
-      expect(consolidateText).not.toContain("not found");
-      expect(consolidateText).toContain("Mode: delete");
-      expect(consolidateText).toContain("Source notes deleted.");
-
-      const consolidatedIdMatch = consolidateText.match(/Consolidated \d+ notes into '([^']+)'/);
-      expect(consolidatedIdMatch).toBeTruthy();
-      const consolidatedId = consolidatedIdMatch![1]!;
-
-      const consolidatedPath = path.join(vaultDir, "notes", `${consolidatedId}.md`);
-      const consolidatedContents = await readFile(consolidatedPath, "utf-8");
-      expect(consolidatedContents).toContain("Merged content from a global note and a project-associated note.");
-
-      // Both source notes must be gone
-      await expect(stat(path.join(vaultDir, "notes", `${globalId}.md`))).rejects.toThrow();
-      await expect(stat(path.join(vaultDir, "notes", `${projectAssociatedId}.md`))).rejects.toThrow();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("rebuilds all embeddings during sync when force=true", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Sync force rebuild note",
-        content: "This note will have its embedding rebuilt by sync force mode.",
-        scope: "global",
-        summary: "Seed note for sync force rebuild test",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-      const embeddingPath = path.join(vaultDir, "embeddings", `${noteId}.json`);
-      const before = await readFile(embeddingPath, "utf-8");
-
-      const response = await callLocalMcpResponse(vaultDir, "sync", { force: true }, embeddingServer.url);
-
-      expect(response.text).toContain("main vault: no remote configured — git sync skipped.");
-      expect(response.text).toContain("main vault: embedded 1 note(s) (force rebuild).");
-      const structured = response.structuredContent;
-      expect(structured?.["action"]).toBe("synced");
-      const vaults = structured?.["vaults"] as Array<Record<string, unknown>>;
-      expect(vaults).toHaveLength(1);
-      expect(vaults[0]?.["vault"]).toBe("main");
-      expect(vaults[0]?.["embedded"]).toBe(1);
-      expect(vaults[0]?.["failed"]).toEqual([]);
-      await expect(stat(embeddingPath)).resolves.toBeDefined();
-      const after = await readFile(embeddingPath, "utf-8");
-      expect(after).not.toBe("");
-      expect(before).not.toBe("");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("keeps sync structured output aligned with SyncResultSchema", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      await callLocalMcp(vaultDir, "remember", {
-        title: "Schema audit sync note",
-        content: "Used to validate SyncResultSchema alignment.",
-        scope: "global",
-        summary: "Seed note for sync schema audit",
-      }, embeddingServer.url);
-
-      const response = await callLocalMcpResponse(vaultDir, "sync", {}, embeddingServer.url);
-
-      // Must parse without throwing — validates gitError optional field is included correctly
-      const parsed = SyncResultSchema.parse(response.structuredContent);
-      expect(parsed.action).toBe("synced");
-      expect(parsed.vaults).toHaveLength(1);
-      expect(parsed.vaults[0]?.vault).toBe("main");
-      // No remote configured so gitError should be absent
-      expect(parsed.vaults[0]?.gitError).toBeUndefined();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("does not create a project vault when updating a main-vault note with cwd pointing to unadopted project", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await initTestRepo(repoDir);
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const rememberText = await callLocalMcp(vaultDir, "remember", {
-        title: "Global note for update test",
-        content: "Original content.",
-        tags: ["update-test"],
-        lifecycle: "permanent",
-        summary: "Create global note for update test",
-        cwd: repoDir,
-        scope: "global",
-      }, embeddingServer.url);
-
-      const noteId = extractRememberedId(rememberText);
-
-      await callLocalMcp(vaultDir, "update", {
-        id: noteId,
-        content: "Updated content.",
-        summary: "Update note content",
-        cwd: repoDir,
-      }, embeddingServer.url);
-
-      // Project vault must NOT have been created as a side effect of passing cwd
-      await expect(stat(path.join(repoDir, ".mnemonic"))).rejects.toThrow();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("does not create a project vault when consolidating main-vault notes with cwd pointing to unadopted project", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-    tempDirs.push(vaultDir, repoDir);
-
-    await execFileAsync("git", ["init"], { cwd: repoDir });
-
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const firstRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Global note A",
-        content: "First global note for consolidation.",
-        tags: ["global-merge"],
-        lifecycle: "temporary",
-        summary: "Create first global note",
-        cwd: repoDir,
-        scope: "global",
-      }, embeddingServer.url);
-      const secondRemember = await callLocalMcp(vaultDir, "remember", {
-        title: "Global note B",
-        content: "Second global note for consolidation.",
-        tags: ["global-merge"],
-        lifecycle: "temporary",
-        summary: "Create second global note",
-        cwd: repoDir,
-        scope: "global",
-      }, embeddingServer.url);
-
-      const firstId = extractRememberedId(firstRemember);
-      const secondId = extractRememberedId(secondRemember);
-
-      await callLocalMcp(vaultDir, "consolidate", {
-        cwd: repoDir,
-        strategy: "execute-merge",
-        mergePlan: {
-          sourceIds: [firstId, secondId],
-          targetTitle: "Merged global note",
-        },
-      }, embeddingServer.url);
-
-      // Project vault must NOT have been created as a side effect
-      await expect(stat(path.join(repoDir, ".mnemonic"))).rejects.toThrow();
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  describe("schema-audit", () => {
-    it("recall temporal mode returns compact history and verbose stats", async () => {
-      const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-      tempDirs.push(vaultDir);
-      const embeddingServer = await startFakeEmbeddingServer();
-
-      try {
-        await initTestVaultRepo(vaultDir);
-
-        await callLocalMcp(vaultDir, "remember", {
-          title: "Temporal recall schema audit test",
-          content: "Testing temporal recall history output.",
-          tags: ["audit", "temporal"],
-          lifecycle: "permanent",
-          scope: "global",
-          summary: "Seed note for temporal recall schema audit",
-        }, { ollamaUrl: embeddingServer.url, disableGit: false });
-
-        const response = await callLocalMcpResponse(vaultDir, "recall", {
-          query: "temporal recall schema audit test",
-          mode: "temporal",
-          verbose: true,
-        }, { ollamaUrl: embeddingServer.url, disableGit: false });
-
-        expect(() => RecallResultSchema.parse(response.structuredContent)).not.toThrow();
-        const parsed = RecallResultSchema.parse(response.structuredContent);
-        expect(parsed.results).toHaveLength(1);
-        expect(parsed.results[0]?.history).toBeDefined();
-        expect(parsed.results[0]?.history?.length).toBeGreaterThan(0);
-        expect(parsed.results[0]?.history?.[0]).toHaveProperty("commitHash");
-        expect(parsed.results[0]?.history?.[0]).toHaveProperty("message");
-        expect(parsed.results[0]?.history?.[0]).toHaveProperty("timestamp");
-        expect(parsed.results[0]?.history?.[0]).toHaveProperty("stats");
-      } finally {
-        await embeddingServer.close();
-      }
-    }, 15000);
-
-    it("recall temporal mode omits history beyond the enrichment cap instead of claiming no history", async () => {
-      const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-      tempDirs.push(vaultDir);
-      const embeddingServer = await startFakeEmbeddingServer();
-
-      try {
-        await initTestVaultRepo(vaultDir);
-
-        for (let i = 1; i <= 6; i++) {
-          await callLocalMcp(vaultDir, "remember", {
-            title: `Temporal recall cap test ${i}`,
-            content: `Testing temporal recall enrichment cap item ${i}.`,
-            tags: ["audit", "temporal", "cap-test"],
-            lifecycle: "permanent",
-            scope: "global",
-            summary: `Seed note ${i} for temporal recall cap audit`,
-          }, { ollamaUrl: embeddingServer.url, disableGit: false });
-        }
-
-        const response = await callLocalMcpResponse(vaultDir, "recall", {
-          query: "temporal recall cap test",
-          mode: "temporal",
-          limit: 6,
-        }, { ollamaUrl: embeddingServer.url, disableGit: false });
-
-        const parsed = RecallResultSchema.parse(response.structuredContent);
-        expect(parsed.results).toHaveLength(6);
-        expect(parsed.results[4]?.history).toBeDefined();
-        expect(parsed.results[5]?.history).toBeUndefined();
-        expect(response.text).not.toContain("no git history found");
-      } finally {
-        await embeddingServer.close();
-      }
-    }, 25000);
-
-    it("plain recall leaves temporal history absent", async () => {
-      const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-      tempDirs.push(vaultDir);
-      const embeddingServer = await startFakeEmbeddingServer();
-
-      try {
-        await initTestVaultRepo(vaultDir);
-
-        await callLocalMcp(vaultDir, "remember", {
-          title: "Plain recall schema audit test",
-          content: "Testing that default recall does not include temporal history.",
-          tags: ["audit", "recall"],
-          lifecycle: "permanent",
-          scope: "global",
-          summary: "Seed note for plain recall schema audit",
-        }, { ollamaUrl: embeddingServer.url, disableGit: false });
-
-        const response = await callLocalMcpResponse(vaultDir, "recall", {
-          query: "plain recall schema audit test",
-        }, { ollamaUrl: embeddingServer.url, disableGit: false });
-
-        const parsed = RecallResultSchema.parse(response.structuredContent);
-        expect(parsed.results).toHaveLength(1);
-        expect(parsed.results[0]?.history).toBeUndefined();
-        expect(response.text).not.toContain("**history:**");
-      } finally {
-        await embeddingServer.close();
-      }
-    }, 15000);
-
-    it("RecallResultSchema accepts provenance and confidence fields in results", async () => {
-      const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-      tempDirs.push(vaultDir);
-      const embeddingServer = await startFakeEmbeddingServer();
-
-      try {
-        await callLocalMcp(vaultDir, "remember", {
-          title: "Schema audit recall test",
-          content: "Testing provenance and confidence in recall results.",
-          tags: ["audit"],
-          lifecycle: "permanent",
-          scope: "global",
-          summary: "Seed note for recall schema audit",
-        }, embeddingServer.url);
-
-        const response = await callLocalMcpResponse(vaultDir, "recall", {
-          query: "schema audit recall test",
-        }, embeddingServer.url);
-
-        expect(() => RecallResultSchema.parse(response.structuredContent)).not.toThrow();
-        const parsed = RecallResultSchema.parse(response.structuredContent);
-        expect(parsed.action).toBe("recalled");
-        expect(parsed.results).toHaveLength(1);
-        // provenance and confidence are optional but should be accepted if present
-        expect(parsed.results[0]).toHaveProperty("id");
-        expect(parsed.results[0]).toHaveProperty("title");
-      } finally {
-        await embeddingServer.close();
-      }
-    }, 15000);
-
-    it("OrientationNoteSchema accepts provenance and confidence optional fields", async () => {
-      const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-      const repoDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-project-"));
-      tempDirs.push(vaultDir, repoDir);
-
-      await initTestRepo(repoDir);
-      const embeddingServer = await startFakeEmbeddingServer();
-
-      try {
-        await callLocalMcp(vaultDir, "remember", {
-          title: "Schema audit orientation test",
-          content: "Testing provenance and confidence in orientation.",
-          tags: ["audit", "orientation"],
-          lifecycle: "permanent",
-          scope: "project",
-          summary: "Seed note for orientation schema audit",
-          cwd: repoDir,
-        }, embeddingServer.url);
-
-        const summary = await callLocalMcpResponse(vaultDir, "project_memory_summary", {
-          cwd: repoDir,
-        }, embeddingServer.url);
-
-        expect(() => ProjectSummaryResultSchema.parse(summary.structuredContent)).not.toThrow();
-        const parsed = ProjectSummaryResultSchema.parse(summary.structuredContent);
-        expect(parsed.action).toBe("project_summary_shown");
-        expect(parsed.orientation).toBeDefined();
-        expect(parsed.orientation.primaryEntry).toBeDefined();
-        expect(parsed.orientation.primaryEntry).toHaveProperty("id");
-        expect(parsed.orientation.primaryEntry).toHaveProperty("title");
-        expect(parsed.orientation.primaryEntry).toHaveProperty("rationale");
-      } finally {
-        await embeddingServer.close();
-      }
-    }, 15000);
-  });
-});
-
-async function callLocalMcp(
-  vaultDir: string,
-  toolName: string,
-  arguments_: Record<string, unknown>,
-  options?: string | { ollamaUrl?: string; disableGit?: boolean },
-): Promise<string> {
-  const response = await callLocalMcpResponse(vaultDir, toolName, arguments_, options);
-  return response.text;
-}
-
-async function callLocalMcpPrompt(vaultDir: string, promptName: string): Promise<string> {
-  const response = await callLocalMcpMethod(vaultDir, 1, "prompts/get", { name: promptName });
-  const messages = response?.result?.messages as Array<{ content?: { text?: string } }> | undefined;
-  const text = messages?.[0]?.content?.text;
-  if (!text) {
-    throw new Error(`Missing prompt response for ${promptName}`);
-  }
-
-  return text;
-}
-
-async function listLocalMcpTools(vaultDir: string): Promise<Array<{ name: string; description?: string }>> {
-  const response = await callLocalMcpMethod(vaultDir, 1, "tools/list", {});
-  const tools = response?.result?.tools as Array<{ name: string; description?: string }> | undefined;
-  if (!tools) {
-    throw new Error("Missing tools/list response");
-  }
-
-  return tools;
-}
-
-async function callLocalMcpResponse(
-  vaultDir: string,
-  toolName: string,
-  arguments_: Record<string, unknown>,
-  options?: string | { ollamaUrl?: string; disableGit?: boolean },
-): Promise<{ text: string; structuredContent?: Record<string, unknown> }> {
-  const resolvedOptions = typeof options === "string" ? { ollamaUrl: options } : options;
-  const response = await callLocalMcpMethod(vaultDir, 1, "tools/call", {
-    name: toolName,
-    arguments: arguments_,
-  }, resolvedOptions);
-  const text = response?.result?.content?.[0]?.text;
-  if (!text) {
-    throw new Error(`Missing tool response for ${toolName}`);
-  }
-
-  return { text, structuredContent: response?.result?.structuredContent as Record<string, unknown> | undefined };
-}
-
-async function callLocalMcpMethod(
-  vaultDir: string,
-  id: number,
-  method: string,
-  params: Record<string, unknown>,
-  options?: { ollamaUrl?: string; disableGit?: boolean },
-): Promise<{ id?: number; result?: Record<string, unknown> }> {
-  const messages = [
-    {
-      jsonrpc: "2.0",
-      id: 0,
-      method: "initialize",
-      params: {
-        protocolVersion: "2024-11-05",
-        capabilities: {},
-        clientInfo: { name: "vitest", version: "1.0" },
-      },
-    },
-    {
-      jsonrpc: "2.0",
-      id,
-      method,
-      params,
-    },
-  ];
-
-  const stdout = await new Promise<string>((resolve, reject) => {
-    const child = spawn("node", [builtEntryPoint], {
-      cwd: repoRoot,
-      env: {
-        ...process.env,
-        DISABLE_GIT: options?.disableGit === false ? "false" : "true",
-        VAULT_PATH: vaultDir,
-        ...(options?.ollamaUrl ? { OLLAMA_URL: options.ollamaUrl } : {}),
-      },
-      stdio: ["pipe", "pipe", "pipe"],
-    });
-
-    let stdoutData = "";
-    let stderrData = "";
-
-    child.stdout.on("data", (chunk) => {
-      stdoutData += chunk.toString();
-    });
-    child.stderr.on("data", (chunk) => {
-      stderrData += chunk.toString();
-    });
-    child.on("error", reject);
-    child.on("close", (code) => {
-      if (code !== 0) {
-        reject(new Error(`MCP script exited with ${code}: ${stderrData}`));
-        return;
-      }
-      resolve(stdoutData);
-    });
-
-    child.stdin.end(messages.map((message) => JSON.stringify(message)).join("\n") + "\n");
-  });
-
-  const lines = stdout.trim().split("\n").filter(Boolean).map((line) => JSON.parse(line) as {
-    id?: number;
-    result?: Record<string, unknown>;
-  });
-
-  return lines.find((line) => line.id === id) ?? {};
-}
-
-function extractRememberedId(text: string): string {
-  const match = text.match(/`([^`]+)`/);
-  if (!match) {
-    throw new Error(`Could not parse remembered id from: ${text}`);
-  }
-
-  return match[1];
-}
-
-async function startFakeEmbeddingServer(): Promise<{ url: string; close: () => Promise<void> }> {
-  const server = http.createServer((req, res) => {
-    if (req.method !== "POST" || req.url !== "/api/embed") {
-      res.writeHead(404).end();
-      return;
-    }
-
-    res.writeHead(200, { "Content-Type": "application/json" });
-    res.end(JSON.stringify({ embeddings: [[0.1, 0.2, 0.3]] }));
-  });
-
-  await new Promise<void>((resolve, reject) => {
-    server.once("error", reject);
-    server.listen(0, "127.0.0.1", () => resolve());
-  });
-
-  const address = server.address();
-  if (!address || typeof address === "string") {
-    throw new Error("Could not determine fake embedding server address");
-  }
-
-  return {
-    url: `http://127.0.0.1:${address.port}`,
-    close: () => new Promise<void>((resolve, reject) => {
-      server.close((err) => err ? reject(err) : resolve());
-    }),
-  };
-}
